% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_that.R
\name{tweet_that}
\alias{tweet_that}
\title{Tweet that}
\usage{
tweet_that(text, url = NULL, via = NULL)
}
\arguments{
\item{text}{the text of the tweet (mandatory)}

\item{url}{optional url}

\item{via}{optional via, with or without the `@`}
}
\value{
open your browser with a tweet ready to be sent
}
\description{
Send a tweet from your R session
}
\examples{
tweet_that(text = "Colin's website is cool", url = "http://colinfay.me", via = "@_colinfay")
}
