% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_widget.R
\name{tweet_widget}
\alias{tweet_widget}
\title{Create a post}
\usage{
tweet_widget(inputValue1, inputValue2)
}
\arguments{
\item{status}{post title}

\item{media}{output directory (default to _posts)}

\item{token}{the output url}

\item{in_reply_to_status_id}{the date of the post}
}
\description{
Create a post
}
